Add game spec, update runner, change start to game
