+++ lib/game.rb
@@ -1,0 +1,20 @@
+require 'rspec'
+require './lib/card'
+require './lib/deck'
+require './lib/player'
+require './lib/turn'
+require './war_or_peace_runner'
+require 'pry'
+
+class Start
+
+  def initialize
+    @player1 = player1
+    @player2 = player2
+    @standard_deck = standard_deck
+  end
+
+  def start_game
+
+  end
+end
