+++ lib/turn.rb
@@ -1,7 +1,7 @@
 require 'pry'
-require './card'
-require './deck'
-require './player'
+require './lib/card'
+require './lib/deck'
+require './lib/player'
 
 class Turn
   attr_reader :player1, :player2, :spoils_of_war, :cards_won, :remove_card, :winner
